[Add]Clean code structure
