+++ lib/game_of_life.rb
@@ -2,6 +2,7 @@
 
 require './game_of_life/game'
 require './game_of_life/menu'
+require './game_of_life/board'
 
 alive = 1
 dead = 0
@@ -11,13 +12,13 @@ print 'Introduce a size of the Game of life grid:'
 size = gets.chomp.to_i
 
 grid = Array.new(size) { Array.new(size) }
-
-game = GameOfLife::Game.new(size, grid, alive, dead)
-menu = GameOfLife::Menu.new(size, game)
-game.reset_game
+board = GameOfLife::Board.new(size, grid, alive, dead)
+game = GameOfLife::Game.new(size, grid)
+menu = GameOfLife::Menu.new(size, game, board)
+board.reset_game
 game.show_grid
 while game.play_again?(answer) != 1
   menu.print_menu
   menu.start_game
-  game.check_grid(stop)
+  board.check_grid(stop)
 end
