+++ lib/game_of_life/menu.rb
@@ -3,9 +3,10 @@
 module GameOfLife
   # Menu
   class Menu
-    def initialize(size, game)
+    def initialize(size, game, board)
       @size = size
       @game = game
+      @board = board
     end
 
     def print_menu
@@ -25,7 +26,7 @@ module GameOfLife
           position_row = gets.chomp.to_i
           puts 'Introduce a column:'
           position_column = gets.chomp.to_i
-          @game.update_grid(position_row, position_column)
+          @board.update_grid(position_row, position_column)
         end
       when 2
         @game.game_over
